+++ src/steps/Microsoft.DotNet.UpgradeAssistant.Steps.Razor/RazorUpdaterStepExtensions.cs
@@ -23,9 +23,11 @@ namespace Microsoft.DotNet.UpgradeAssistant
             services
             .AddUpgradeStep<RazorUpdaterStep>()
             .AddTransient<IUpdater<RazorCodeDocument>, RazorSourceUpdater>()
+            .AddTransient<IUpdater<RazorCodeDocument>, RazorHelperUpdater>()
             .AddTransient<ITextMatcher, DefaultTextMatcher>()
-            .AddTransient<ITextReplacer, RazorTextReplacer>()
+            .AddTransient<IMappedTextReplacer, RazorMappedTextReplacer>()
             .AddTransient<IDiffer, Differ>()
-            .AddTransient<IChunker, CharacterChunker>();
+            .AddTransient<IChunker, CharacterChunker>()
+            .AddTransient<IHelperMatcher, HelperMatcher>();
     }
 }