+++ tests/steps/Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests/RazorSourceUpdaterTests.cs
@@ -26,11 +26,11 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
         [Fact]
         public void CtorNegativeTests()
         {
-            using var mock = GetMock("RazorUpdaterStepViews/Test.csproj", Array.Empty<LocationLookup[]>());
+            using var mock = GetMock("RazorSourceUpdaterStepViews/Test.csproj", Array.Empty<LocationLookup[]>());
             var analyzers = mock.Container.Resolve<IEnumerable<DiagnosticAnalyzer>>();
             var textMatcher = mock.Container.Resolve<ITextMatcher>();
             var codeFixProviders = mock.Container.Resolve<IEnumerable<CodeFixProvider>>();
-            var textReplacer = mock.Mock<ITextReplacer>();
+            var textReplacer = mock.Mock<IMappedTextReplacer>();
             var logger = mock.Mock<ILogger<RazorSourceUpdater>>();
 
             Assert.Throws<ArgumentNullException>("analyzers", () => new RazorSourceUpdater(null!, codeFixProviders, ImmutableArray<AdditionalText>.Empty, textMatcher, textReplacer.Object, logger.Object));
@@ -44,7 +44,7 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
         [Fact]
         public void PropertyTests()
         {
-            using var mock = GetMock("RazorUpdaterStepViews/Test.csproj", Array.Empty<LocationLookup[]>());
+            using var mock = GetMock("RazorSourceUpdaterStepViews/Test.csproj", Array.Empty<LocationLookup[]>());
             var updater = mock.Create<RazorSourceUpdater>();
 
             Assert.Equal("Microsoft.DotNet.UpgradeAssistant.Steps.Razor.RazorSourceUpdater", updater.Id);
@@ -68,7 +68,7 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
         public async Task IsApplicableTests(LocationLookup[][] diagnosticLocations, string[] expectedApplicableFilePaths)
         {
             // Arrange
-            using var mock = GetMock("RazorUpdaterStepViews/Test.csproj", diagnosticLocations);
+            using var mock = GetMock("RazorSourceUpdaterStepViews/Test.csproj", diagnosticLocations);
             var razorDocs = await GetRazorCodeDocumentsAsync(mock).ConfigureAwait(true);
             var context = mock.Mock<IUpgradeContext>();
             var updater = mock.Create<RazorSourceUpdater>();
@@ -106,9 +106,9 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
                 {
                     new[]
                     {
-                        new[] { new LocationLookup("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs", "Model[0]") },
+                        new[] { new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs", "Model[0]") },
                     },
-                    new[] { GetFullPath("RazorUpdaterStepViews\\TestViews\\View.cshtml") },
+                    new[] { GetFullPath("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml") },
                 },
 
                 // Diagnostic mapped to shared Razor file
@@ -116,9 +116,9 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
                 {
                     new[]
                     {
-                        new[] { new LocationLookup("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs", "using Microsoft.AspNetCore.Mvc;") },
+                        new[] { new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs", "using Microsoft.AspNetCore.Mvc;") },
                     },
-                    new[] { GetFullPath("RazorUpdaterStepViews\\_ViewImports.cshtml") },
+                    new[] { GetFullPath("RazorSourceUpdaterStepViews\\_ViewImports.cshtml") },
                 },
 
                 // Diagnostic in unmapped portions of generated files
@@ -128,13 +128,13 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
                     {
                         new[]
                         {
-                            new LocationLookup("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs", "assembly: global::Microsoft.AspNetCore"),
-                            new LocationLookup("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs", "WriteLiteral(\"    <div>\\r\\n        <p>\")"),
+                            new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs", "assembly: global::Microsoft.AspNetCore"),
+                            new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs", "WriteLiteral(\"    <div>\\r\\n        <p>\")"),
                         },
                     },
                     new[]
                     {
-                        GetFullPath("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs"),
+                        GetFullPath("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs"),
                     },
                 },
 
@@ -145,21 +145,21 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
                     {
                         new[]
                         {
-                            new LocationLookup("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs", "using Microsoft.AspNetCore.Mvc;"),
-                            new LocationLookup("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs", "Model[0]"),
+                            new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs", "using Microsoft.AspNetCore.Mvc;"),
+                            new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs", "Model[0]"),
                         },
                         new[]
                         {
-                            new LocationLookup("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs", "Model[1]"),
-                            new LocationLookup("RazorUpdaterStepViews\\TestViews\\Simple.cshtml.cs", "using Microsoft.AspNetCore.Mvc;"),
-                            new LocationLookup("RazorUpdaterStepViews\\TestViews\\Simple.cshtml.cs", "DateTime.Now.ToString()"),
+                            new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs", "Model[1]"),
+                            new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\Simple.cshtml.cs", "using Microsoft.AspNetCore.Mvc;"),
+                            new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\Simple.cshtml.cs", "DateTime.Now.ToString()"),
                         }
                     },
                     new[]
                     {
-                        GetFullPath("RazorUpdaterStepViews\\_ViewImports.cshtml"),
-                        GetFullPath("RazorUpdaterStepViews\\TestViews\\View.cshtml"),
-                        GetFullPath("RazorUpdaterStepViews\\TestViews\\Simple.cshtml"),
+                        GetFullPath("RazorSourceUpdaterStepViews\\_ViewImports.cshtml"),
+                        GetFullPath("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml"),
+                        GetFullPath("RazorSourceUpdaterStepViews\\TestViews\\Simple.cshtml"),
                     },
                 }
             };
@@ -176,17 +176,17 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
 
         [Theory]
         [MemberData(nameof(ApplyData))]
-        public async Task ApplyTests(LocationLookup[][] diagnosticLocations, string[] expectedUpdatedFiles, TextReplacement[] expectedReplacements)
+        public async Task ApplyTests(LocationLookup[][] diagnosticLocations, string[] expectedUpdatedFiles, MappedTextReplacement[] expectedReplacements)
         {
             // Arrange
-            using var mock = GetMock("RazorUpdaterStepViews/Test.csproj", diagnosticLocations);
+            using var mock = GetMock("RazorSourceUpdaterStepViews/Test.csproj", diagnosticLocations);
             var razorDocs = await GetRazorCodeDocumentsAsync(mock).ConfigureAwait(true);
             var context = mock.Mock<IUpgradeContext>();
             var updater = mock.Create<RazorSourceUpdater>();
-            var replacements = new List<TextReplacement>();
-            var textReplacer = mock.Mock<ITextReplacer>();
-            textReplacer.Setup(r => r.ApplyTextReplacements(It.IsAny<IEnumerable<TextReplacement>>()))
-                .Callback<IEnumerable<TextReplacement>>(newReplacements => replacements.AddRange(newReplacements));
+            var replacements = new List<MappedTextReplacement>();
+            var textReplacer = mock.Mock<IMappedTextReplacer>();
+            textReplacer.Setup(r => r.ApplyTextReplacements(It.IsAny<IEnumerable<MappedTextReplacement>>()))
+                .Callback<IEnumerable<MappedTextReplacement>>(newReplacements => replacements.AddRange(newReplacements));
 
             // Act
             var result = (FileUpdaterResult)await updater.ApplyAsync(context.Object, ImmutableArray.CreateRange(razorDocs), CancellationToken.None).ConfigureAwait(true);
@@ -197,7 +197,7 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
             Assert.Collection(result.FilePaths.OrderBy(f => f), expectedUpdatedFiles.OrderBy(f => f).Select<string, Action<string>>(expected => actual => Assert.Equal(expected, actual)).ToArray());
             Assert.True(resultWithoutDocs.Result);
             Assert.Empty(resultWithoutDocs.FilePaths);
-            Assert.Collection(replacements, expectedReplacements.Select<TextReplacement, Action<TextReplacement>>(e => a =>
+            Assert.Collection(replacements, expectedReplacements.Select<MappedTextReplacement, Action<MappedTextReplacement>>(e => a =>
             {
                 if (string.IsNullOrEmpty(e.OriginalText.ToString()) && string.IsNullOrEmpty(e.NewText.ToString()))
                 {
@@ -219,7 +219,7 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
                 {
                     Array.Empty<LocationLookup[]>(),
                     Array.Empty<string>(),
-                    Array.Empty<TextReplacement>(),
+                    Array.Empty<MappedTextReplacement>(),
                 },
 
                 // Diagnostics in non-Razor files
@@ -227,7 +227,7 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
                 {
                     new[] { new[] { new LocationLookup("Foo.cs", null, 10, 15) } },
                     Array.Empty<string>(),
-                    Array.Empty<TextReplacement>(),
+                    Array.Empty<MappedTextReplacement>(),
                 },
 
                 // Diagnostic in Razor file
@@ -235,10 +235,10 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
                 {
                     new[]
                     {
-                        new[] { new LocationLookup("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs", "Model[0]") },
+                        new[] { new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs", "Model[0]") },
                     },
-                    new[] { GetFullPath("RazorUpdaterStepViews\\TestViews\\View.cshtml") },
-                    new[] { new TextReplacement("      Write(Model[0]);\r\n", "      Write(Model[0] /* Test! */);\r\n", GetFullPath("RazorUpdaterStepViews\\TestViews\\View.cshtml"), 6) }
+                    new[] { GetFullPath("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml") },
+                    new[] { new MappedTextReplacement("      Write(Model[0]);\r\n", "      Write(Model[0] /* Test! */);\r\n", GetFullPath("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml"), 6) }
                 },
 
                 // Diagnostic mapped to shared Razor file
@@ -246,10 +246,10 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
                 {
                     new[]
                     {
-                        new[] { new LocationLookup("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs", "using Microsoft.AspNetCore.Mvc;") },
+                        new[] { new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs", "using Microsoft.AspNetCore.Mvc;") },
                     },
-                    new[] { GetFullPath("RazorUpdaterStepViews\\_ViewImports.cshtml") },
-                    new[] { new TextReplacement("using Microsoft.AspNetCore.Mvc;\r\n", "using Microsoft.AspNetCore.Mvc; /* Test! */\r\n", GetFullPath("RazorUpdaterStepViews\\_ViewImports.cshtml"), 1) }
+                    new[] { GetFullPath("RazorSourceUpdaterStepViews\\_ViewImports.cshtml") },
+                    new[] { new MappedTextReplacement("using Microsoft.AspNetCore.Mvc;\r\n", "using Microsoft.AspNetCore.Mvc; /* Test! */\r\n", GetFullPath("RazorSourceUpdaterStepViews\\_ViewImports.cshtml"), 1) }
                 },
 
                 // Diagnostic in unmapped portions of generated files
@@ -259,16 +259,16 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
                     {
                         new[]
                         {
-                            new LocationLookup("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs", "assembly: global::Microsoft.AspNetCore"),
-                            new LocationLookup("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs", "WriteLiteral(\"    <div>\\r\\n        <p>\")"),
+                            new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs", "assembly: global::Microsoft.AspNetCore"),
+                            new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs", "WriteLiteral(\"    <div>\\r\\n        <p>\")"),
                         },
                     },
-                    new[] { GetFullPath("RazorUpdaterStepViews\\TestViews\\View.cshtml") },
+                    new[] { GetFullPath("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml") },
                     new[]
                     {
                         // The first one *does* generate a replacement because it represents text being prepended to the beginning of the source file
                         // Don't check the actual text, though, since it will include file path-specific values that will change
-                        new TextReplacement(string.Empty, string.Empty, GetFullPath("RazorUpdaterStepViews\\TestViews\\View.cshtml"), 0),
+                        new MappedTextReplacement(string.Empty, string.Empty, GetFullPath("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml"), 0),
                     }
                 },
 
@@ -279,28 +279,28 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor.Tests
                     {
                         new[]
                         {
-                            new LocationLookup("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs", "using Microsoft.AspNetCore.Mvc;"),
-                            new LocationLookup("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs", "Model[0]"),
+                            new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs", "using Microsoft.AspNetCore.Mvc;"),
+                            new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs", "Model[0]"),
                         },
                         new[]
                         {
-                            new LocationLookup("RazorUpdaterStepViews\\TestViews\\View.cshtml.cs", "Model[1]"),
-                            new LocationLookup("RazorUpdaterStepViews\\TestViews\\Simple.cshtml.cs", "using Microsoft.AspNetCore.Mvc;"),
-                            new LocationLookup("RazorUpdaterStepViews\\TestViews\\Simple.cshtml.cs", "DateTime.Now.ToString()"),
+                            new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml.cs", "Model[1]"),
+                            new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\Simple.cshtml.cs", "using Microsoft.AspNetCore.Mvc;"),
+                            new LocationLookup("RazorSourceUpdaterStepViews\\TestViews\\Simple.cshtml.cs", "DateTime.Now.ToString()"),
                         }
                     },
                     new[]
                     {
-                        GetFullPath("RazorUpdaterStepViews\\_ViewImports.cshtml"),
-                        GetFullPath("RazorUpdaterStepViews\\TestViews\\View.cshtml"),
-                        GetFullPath("RazorUpdaterStepViews\\TestViews\\Simple.cshtml"),
+                        GetFullPath("RazorSourceUpdaterStepViews\\_ViewImports.cshtml"),
+                        GetFullPath("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml"),
+                        GetFullPath("RazorSourceUpdaterStepViews\\TestViews\\Simple.cshtml"),
                     },
                     new[]
                     {
-                        new TextReplacement("using Microsoft.AspNetCore.Mvc;\r\n", "using Microsoft.AspNetCore.Mvc; /* Test! */\r\n", GetFullPath("RazorUpdaterStepViews\\_ViewImports.cshtml"), 1),
-                        new TextReplacement(" Write(DateTime.Now.ToString());\r\n", " Write(DateTime.Now.ToString() /* Test! */);\r\n", GetFullPath("RazorUpdaterStepViews\\TestViews\\Simple.cshtml"), 1),
-                        new TextReplacement("      Write(Model[0]);\r\n", "      Write(Model[0] /* Test! */);\r\n", GetFullPath("RazorUpdaterStepViews\\TestViews\\View.cshtml"), 6),
-                        new TextReplacement("      Write(Model[1]);\r\n", "      Write(Model[1] /* Test! */);\r\n", GetFullPath("RazorUpdaterStepViews\\TestViews\\View.cshtml"), 18),
+                        new MappedTextReplacement("using Microsoft.AspNetCore.Mvc;\r\n", "using Microsoft.AspNetCore.Mvc; /* Test! */\r\n", GetFullPath("RazorSourceUpdaterStepViews\\_ViewImports.cshtml"), 1),
+                        new MappedTextReplacement(" Write(DateTime.Now.ToString());\r\n", " Write(DateTime.Now.ToString() /* Test! */);\r\n", GetFullPath("RazorSourceUpdaterStepViews\\TestViews\\Simple.cshtml"), 1),
+                        new MappedTextReplacement("      Write(Model[0]);\r\n", "      Write(Model[0] /* Test! */);\r\n", GetFullPath("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml"), 6),
+                        new MappedTextReplacement("      Write(Model[1]);\r\n", "      Write(Model[1] /* Test! */);\r\n", GetFullPath("RazorSourceUpdaterStepViews\\TestViews\\View.cshtml"), 18),
                     }
                 }
             };